+++ Source/SelfService/Web/microservice/microserviceView/microserviceView.tsx
@@ -18,7 +18,6 @@ import { ContainerHealthStatus } from '../microserviceStatus';
 import { useTerminalAvailable } from './terminal/useTerminal';
 import { View as Terminal } from './terminal/View';
 import { View as Management } from './management/View';
-import { useRuntimeManagementAvailable } from './management/useRuntimeManagement';
 
 type MicroserviceViewProps = {
     application: HttpResponseApplication;
@@ -110,6 +109,14 @@ export const MicroserviceView = ({ application, microserviceId, environment, pod
                 microserviceId={microserviceId}
                 data={podsData}
             />
+        },
+        {
+            label: 'Management',
+            render: () => <Management
+                applicationId={applicationId}
+                environment={environment}
+                microserviceId={microserviceId}
+            />
         }
     ];
 
@@ -124,21 +131,6 @@ export const MicroserviceView = ({ application, microserviceId, environment, pod
                     microserviceId={microserviceId}
                 />
             },
-        );
-    }
-
-    // const managementAvailable = useRuntimeManagementAvailable(applicationId, environment, microserviceId);
-    // console.log('management available', managementAvailable);
-    if (true) {
-        tabs.push(
-            {
-                label: 'Management',
-                render: () => <Management
-                    applicationId={applicationId}
-                    environment={environment}
-                    microserviceId={microserviceId}
-                />
-            }
         );
     }
